requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
transformers
git+https://github.com/openai/CLIP.git
timm